Move RecordsHistorian.purgeExpiredRecords outside ledger txn
